requirements.txt
Flask
twilio
requests
python-dotenv
elevenlabs
